Correctly interpolate values of all types
The previous code interpolated list of character codes and lists of
single letter atoms (codes and chars) as list literals.  That's fine
for many circumstances, but in the context of *string* interpolation,
we want to treat those two types as strings.

I hoped that ~s would work across the board but it throws an exception
with terms, numbers and variables.  ~p is semantically the right
choice because it lets compound terms decide how they want to be
displayed (via portray/1 hook).  I settled on using ~p for everything
except chars and codes, which both use ~s.